Move to plugins-core 0.7.2
- also changes extraFilesDirectory -> extraFilesDirectory
