+++ src/main/java/com/google/cloud/tools/gradle/appengine/standard/DevAppServerStopTask.java
@@ -17,10 +17,10 @@
 
 package com.google.cloud.tools.gradle.appengine.standard;
 
-import com.google.cloud.tools.appengine.api.AppEngineException;
-import com.google.cloud.tools.appengine.api.devserver.AppEngineDevServer;
-import com.google.cloud.tools.appengine.cloudsdk.CloudSdkNotFoundException;
-import com.google.cloud.tools.appengine.cloudsdk.LocalRun;
+import com.google.cloud.tools.appengine.AppEngineException;
+import com.google.cloud.tools.appengine.operations.DevServer;
+import com.google.cloud.tools.appengine.operations.LocalRun;
+import com.google.cloud.tools.appengine.operations.cloudsdk.CloudSdkNotFoundException;
 import com.google.cloud.tools.gradle.appengine.core.CloudSdkOperations;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.TaskAction;
@@ -43,7 +43,7 @@ public class DevAppServerStopTask extends DefaultTask {
   /** Task entrypoint : Stop the dev appserver (get StopConfiguration from helper). */
   @TaskAction
   public void stopAction() throws CloudSdkNotFoundException {
-    AppEngineDevServer server =
+    DevServer server =
         serverHelper.getAppServer(
             localRun, runConfig, CloudSdkOperations.getDefaultHandler(getLogger()));
     try {
